refactor: colocate motionCreated helpers with handler
