+++ src/handlers/motions/motionCreated/index.ts
@@ -1,0 +1,1 @@
+export { default } from './motionCreated';
